remove mkl baxed CSR creation

+++ cpp/main.cpp
@@ -170,14 +170,14 @@ int main(int argc, char* argv[]) {
   ;
   auto t = start_clock();
   if (file_format == 0){
-     FileReader<uint64_t,float>::ubyte_read(
+     FileReader<int,float>::ubyte_read(
         input_path, data_matrix_ptr.get(), data_set_size, dimension,
         grid.get()->rank_in_col, grid.get()->col_world_size);
     }else if (file_format == 1) {
-      FileReader<uint64_t,float>::fvecs_read(input_path, data_matrix_ptr.get(), data_set_size, dimension,
+      FileReader<int,float>::fvecs_read(input_path, data_matrix_ptr.get(), data_set_size, dimension,
                                     grid.get()->rank_in_col, grid.get()->col_world_size);
     }else if (file_format == 2) {
-      FileReader<uint64_t,float>::read_fbin_with_MPI(input_path, data_matrix_ptr.get(), data_set_size, dimension,
+      FileReader<uint64_t ,float>::read_fbin_with_MPI(input_path, data_matrix_ptr.get(), data_set_size, dimension,
                                     grid.get());
 
     }
@@ -188,7 +188,7 @@ int main(int argc, char* argv[]) {
   std::cout << "calling data loading completed "<<rank<<" "<<std::endl;
 
   std::cout << "calling KNNGHandler rank "<<rank<<" with input matrix:  "<<data_matrix_ptr->size()<<"*"<<(*data_matrix_ptr)[0].size()<<std::endl;
-  auto knng_handler = unique_ptr<KNNGHandler<uint64_t ,float>>(new KNNGHandler<uint64_t,float>(ntrees,  tree_depth,  tree_depth_ratio,
+  auto knng_handler = unique_ptr<KNNGHandler<int ,float>>(new KNNGHandler<int,float>(ntrees,  tree_depth,  tree_depth_ratio,
                                                                                        local_tree_offset,  data_set_size,
                                                                                       data_matrix_ptr.get()->size(),
                                                                                       dimension,  grid.get()));
